Move `FormFieldEditor` to `FormCreator`
