+++ frontend/src/metabase/actions/containers/ActionCreator/FormCreator/FormFieldEditor/FormFieldEditor.tsx
@@ -4,6 +4,7 @@ import { t } from "ttag";
 import Radio from "metabase/core/components/Radio";
 import { isNotNull } from "metabase/core/utils/types";
 
+import ActionFormFieldWidget from "metabase/actions/components/ActionFormFieldWidget";
 import { getFieldTypes, getInputTypes } from "metabase/actions/constants";
 import { inputTypeHasOptions } from "metabase/actions/utils";
 
@@ -14,8 +15,7 @@ import type {
 } from "metabase-types/api";
 import type { ActionFormFieldProps } from "metabase/actions/types";
 
-import { FieldSettingsButtons } from "../../containers/ActionCreator/FormCreator/FieldSettingsButtons";
-import ActionFormFieldWidget from "../ActionFormFieldWidget";
+import { FieldSettingsButtons } from "../FieldSettingsButtons";
 
 import {
   Column,
