+++ frontend/src/metabase/actions/containers/ActionCreator/FormCreator/FormFieldEditor/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FormFieldEditor";
